requirements.txt
torch
transformers
Pillow
langchain
langchain-community
faiss-cpu
numpy
scikit-learn
sentence-transformers
